requirements.txt
slack-bolt
anthropic
python-dotenv